Сднелана основа main
